refactor: extract charts/tables/html/pdf from phi_scan/report package
Split the remaining report monolith into sibling modules:
  - charts.py: _TrendDataPoint, chart builders, audit sanitization boundary
  - tables.py: _build_compliance_matrix_rows
  - html.py:   _HTML_TEMPLATE, _build_html_context, generate_html_report
  - pdf.py:    PDF layout helpers, _PdfTableColumn, generate_pdf_report

phi_scan/report/__init__.py is now a thin compatibility shim that
re-exports the public API (generate_html_report, generate_pdf_report)
plus the private symbols historically imported by tests
(_TrendDataPoint, _GENERAL_REMEDIATION_CHECKLIST, _render_chart_to_bytes,
_build_*_chart).
