+++ phi_scan/report/__init__.py
@@ -22,1274 +22,32 @@ Usage::
 
 from __future__ import annotations
 
-import io
-import textwrap
-import types
-from collections import Counter
-from dataclasses import dataclass
-from datetime import UTC, datetime
-from operator import itemgetter
-from pathlib import Path
-from typing import TYPE_CHECKING
-
-from phi_scan import __version__
-from phi_scan.constants import (
-    CODE_CONTEXT_REDACTED_VALUE,
-    HIPAA_REMEDIATION_GUIDANCE,
-    SEVERITY_RANK,
-    PhiCategory,
-    SeverityLevel,
-)
-from phi_scan.logging_config import get_logger
 from phi_scan.report._shared import (
-    _CHART_HEIGHT_CATEGORY_INCHES,
-    _CHART_HEIGHT_FILES_INCHES,
-    _CHART_HEIGHT_PIE_INCHES,
-    _CHART_HEIGHT_TREND_INCHES,
-    _CHART_WIDTH_INCHES,
     _GENERAL_REMEDIATION_CHECKLIST,
-    _HEX_CRITICAL_RED,
-    _HEX_HIGH_ORANGE,
-    _HEX_LOW_GREEN,
-    _configure_matplotlib_backend,
-    _convert_hex_to_rgb,
-    _get_risk_colour,
-    _MatplotlibFigure,
-    _render_chart_to_base64,
     _render_chart_to_bytes,
 )
+from phi_scan.report.charts import (
+    _build_category_chart,
+    _build_severity_chart,
+    _build_top_files_chart,
+    _build_trend_chart,
+    _TrendDataPoint,
+)
+from phi_scan.report.html import generate_html_report
+from phi_scan.report.pdf import generate_pdf_report
 
-if TYPE_CHECKING:
-    from collections.abc import Mapping
-
-    from phi_scan.compliance import ComplianceControl
-    from phi_scan.models import ScanResult
-
+# Private symbols listed in __all__ so tests can import them directly from
+# `phi_scan.report` without reaching into submodules. The canonical home for
+# each is the corresponding sibling module; this package root is a
+# compatibility / re-export layer only.
 __all__ = [
+    "_GENERAL_REMEDIATION_CHECKLIST",
+    "_TrendDataPoint",
+    "_build_category_chart",
+    "_build_severity_chart",
+    "_build_top_files_chart",
+    "_build_trend_chart",
+    "_render_chart_to_bytes",
     "generate_html_report",
     "generate_pdf_report",
 ]
-
-_logger = get_logger("report")
-
-# ---------------------------------------------------------------------------
-# Layout and colour constants
-# ---------------------------------------------------------------------------
-
-# Severity row colours for PDF table fill (light tint variants)
-_COLOUR_HIGH_TINT: str = "FADBD8"
-_COLOUR_MEDIUM_TINT: str = "FDEBD0"
-_COLOUR_LOW_TINT: str = "D5F5E3"
-_COLOUR_INFO_TINT: str = "EBF5FB"
-_COLOUR_WHITE: str = "FFFFFF"
-
-# PDF page dimensions (A4 portrait, mm)
-_PAGE_WIDTH_MM: float = 210.0
-_PAGE_HEIGHT_MM: float = 297.0
-_MARGIN_MM: float = 15.0
-_PAGE_SIDE_MARGIN_COUNT: int = 2  # one left margin + one right margin
-_CONTENT_WIDTH_MM: float = _PAGE_WIDTH_MM - _PAGE_SIDE_MARGIN_COUNT * _MARGIN_MM
-
-# PDF font sizes
-_FONT_TITLE: int = 22
-_FONT_HEADING: int = 14
-_FONT_SUBHEADING: int = 11
-_FONT_BODY: int = 9
-_FONT_SMALL: int = 8
-_FONT_TABLE_HEADER: int = 8
-_FONT_TABLE_BODY: int = 7
-
-# Maximum code-context characters shown per finding in reports
-_MAX_CONTEXT_CHARS: int = 200
-
-# Top-N files shown in the files-with-most-findings chart
-_TOP_FILES_COUNT: int = 10
-
-# Chart file-label truncation threshold (chars before "..." prefix is added)
-_MAX_LABEL_CHARS: int = 40
-
-# Chart colours (matplotlib format: "#RRGGBB"). Shared colours are derived from
-# the canonical hex constants above so PDF and chart renderers stay in sync.
-_CHART_COLOUR_ACTIVE: str = f"#{_HEX_CRITICAL_RED}"  # category bar with findings present
-_CHART_COLOUR_ZERO_FINDINGS_BAR: str = "#BDC3C7"  # category bar with zero findings
-_CHART_COLOUR_UNKNOWN_SEVERITY: str = "#95A5A6"  # severity donut fallback for unrecognised levels
-_CHART_COLOUR_BLUE: str = "#2980B9"  # trend line and top-files bars
-_CHART_COLOUR_SEVERITY_HIGH: str = f"#{_HEX_CRITICAL_RED}"
-_CHART_COLOUR_SEVERITY_MEDIUM: str = f"#{_HEX_HIGH_ORANGE}"
-_CHART_COLOUR_SEVERITY_LOW: str = f"#{_HEX_LOW_GREEN}"
-_CHART_COLOUR_SEVERITY_INFO: str = "#3498DB"
-
-# Trend-line chart style constants
-_CHART_LINE_WIDTH: int = 2
-_CHART_FILL_ALPHA: float = 0.15
-_CHART_MARKER_STYLE: str = "o"
-
-# Chart annotation and label style
-_CHART_BAR_LABEL_PADDING: int = 3
-_CHART_BAR_LABEL_FONT_SIZE: int = 8
-_CHART_NO_DATA_FONT_SIZE: int = 14
-_CHART_NO_HISTORY_FONT_SIZE: int = 12
-_CHART_PIE_START_ANGLE: int = 90
-_CHART_GRID_ALPHA: float = 0.5
-_CHART_DONUT_WIDTH: float = 0.5
-
-# Chart placeholder text for empty-data states
-_CHART_NO_FINDINGS_LABEL: str = "(no findings)"
-_CHART_NO_FINDINGS_TEXT: str = "No findings"
-_CHART_NO_HISTORY_TEXT: str = "No audit history available"
-
-_PDF_ROW_HEIGHT: float = 5.5
-_PDF_HEADER_HEIGHT: float = 7.0
-
-# Height reserved at the bottom of each PDF page for the footer area.
-# The page-break threshold is derived from this so the two values stay in sync.
-_PDF_FOOTER_HEIGHT_MM: float = 25.0
-# Page-break threshold for findings table (stop before footer area)
-_PDF_PAGE_BREAK_Y_MM: float = _PAGE_HEIGHT_MM - _PDF_FOOTER_HEIGHT_MM
-
-# Audit row column keys consumed by the trend chart sanitization boundary.
-# Only these two columns are ever read from audit rows in report.py.
-# The audit schema stores SHA-256 digests for all path-based identifiers
-# and never persists raw PHI values, so both columns are guaranteed PHI-free:
-#   timestamp      — ISO 8601 datetime string; no PHI
-#   findings_count — INTEGER aggregate count; no PHI
-_AUDIT_ROW_KEY_TIMESTAMP: str = "timestamp"
-_AUDIT_ROW_KEY_FINDINGS_COUNT: str = "findings_count"
-# Frozenset of the only audit row keys permitted to enter the trend chart
-# sanitization boundary. Used at runtime to restrict each row dict to exactly
-# these two keys before any value is extracted — no other column can reach
-# chart rendering regardless of schema changes or unexpected extra fields.
-_AUDIT_ROW_ALLOWED_KEYS: frozenset[str] = frozenset(
-    {_AUDIT_ROW_KEY_TIMESTAMP, _AUDIT_ROW_KEY_FINDINGS_COUNT}
-)
-
-
-# ---------------------------------------------------------------------------
-# PDF table column definitions
-# ---------------------------------------------------------------------------
-
-
-@dataclass(frozen=True)
-class _PdfTableColumn:
-    """Header label and width in mm for a single PDF findings table column."""
-
-    header: str
-    width_mm: float
-
-
-_PDF_TABLE_COLUMNS: tuple[_PdfTableColumn, ...] = (
-    _PdfTableColumn(header="#", width_mm=8.0),
-    _PdfTableColumn(header="File Path", width_mm=52.0),
-    _PdfTableColumn(header="Line", width_mm=10.0),
-    _PdfTableColumn(header="Entity Type", width_mm=28.0),
-    _PdfTableColumn(header="HIPAA Category", width_mm=26.0),
-    _PdfTableColumn(header="Severity", width_mm=16.0),
-    _PdfTableColumn(header="Conf.", width_mm=14.0),
-    _PdfTableColumn(header="Layer", width_mm=16.0),
-)
-
-
-@dataclass(frozen=True)
-class _TrendDataPoint:
-    """One sanitized (scan_date, findings_count) pair extracted from a single audit row."""
-
-    scan_date: datetime
-    findings_count: int
-
-
-# ---------------------------------------------------------------------------
-# Colour helpers
-# ---------------------------------------------------------------------------
-
-# Named RGB tuples for direct fpdf2 set_fill_color / set_text_color calls.
-# fpdf2 core font rendering requires integer (R, G, B) tuples — hex strings cannot be
-# used directly with these methods, so we maintain a parallel set of RGB constants.
-_COLOUR_NAVY_RGB: tuple[int, int, int] = (44, 62, 80)  # dark header and body text
-_COLOUR_WHITE_RGB: tuple[int, int, int] = (255, 255, 255)  # white text on dark fill
-_COLOUR_LIGHT_SILVER_RGB: tuple[int, int, int] = (200, 210, 220)  # cover subtitle text
-_COLOUR_MUTED_GRAY_RGB: tuple[int, int, int] = (127, 140, 141)  # footer and small text
-_COLOUR_CLEAN_GREEN_RGB: tuple[int, int, int] = (39, 174, 96)  # clean / no-findings label
-_COLOUR_DARK_BLUE_RGB: tuple[int, int, int] = (52, 73, 94)  # table header fill
-
-# Y-axis positions on the PDF cover page (millimetres from top of page).
-_PDF_COVER_TITLE_Y_MM: float = 18.0  # vertical position for the report title
-_PDF_COVER_RISK_Y_MM: float = 72.0  # vertical position for the risk-level label
-
-
-# Map of Unicode characters that Helvetica (latin-1) cannot encode → ASCII fallbacks.
-# fpdf2 raises FPDFUnicodeEncodingException for any character outside latin-1 when
-# using the built-in core fonts.  We replace the most common typographic characters
-# before passing text to any fpdf cell/multi_cell call.
-_PDF_CHAR_REPLACEMENTS: types.MappingProxyType[str, str] = types.MappingProxyType(
-    {
-        "\u2014": "-",  # em dash
-        "\u2013": "-",  # en dash
-        "\u2026": "...",  # ellipsis
-        "\u2022": "-",  # bullet
-        "\u00a7": "SS",  # section sign (§)
-        "\u00a9": "(c)",  # copyright
-        "\u00ae": "(R)",  # registered
-        "\u2019": "'",  # right single quote
-        "\u2018": "'",  # left single quote
-        "\u201c": '"',  # left double quote
-        "\u201d": '"',  # right double quote
-    }
-)
-
-
-def _encode_pdf_text_as_latin1(text: str) -> str:
-    """Replace Unicode characters with ASCII fallbacks, then encode to latin-1 for fpdf2."""
-    for char, replacement in _PDF_CHAR_REPLACEMENTS.items():
-        text = text.replace(char, replacement)
-    # Final fallback: drop any remaining non-latin-1 characters
-    return text.encode("latin-1", errors="replace").decode("latin-1")
-
-
-def _get_severity_row_colour(severity: SeverityLevel) -> str:
-    """Return the light-tint hex colour for a severity-level table row."""
-    tint_map: dict[SeverityLevel, str] = {
-        SeverityLevel.HIGH: _COLOUR_HIGH_TINT,
-        SeverityLevel.MEDIUM: _COLOUR_MEDIUM_TINT,
-        SeverityLevel.LOW: _COLOUR_LOW_TINT,
-        SeverityLevel.INFO: _COLOUR_INFO_TINT,
-    }
-    return tint_map.get(severity, _COLOUR_WHITE)
-
-
-def _truncate_code_context(context: str) -> str:
-    """Truncate code context for report rendering, stripping unredacted content.
-
-    The scanner layer is expected to substitute detected PHI values with
-    CODE_CONTEXT_REDACTED_VALUE before populating ScanFinding.code_context.  When the
-    marker is absent from a non-empty context, the content may contain raw PHI —
-    the entire context is replaced with CODE_CONTEXT_REDACTED_VALUE and a warning is logged
-    so that any scanner redaction regression surfaces immediately without
-    allowing PHI to escape into report output.
-
-    Args:
-        context: Raw code_context field from a ScanFinding.
-
-    Returns:
-        Truncated context string, or CODE_CONTEXT_REDACTED_VALUE if redaction cannot be confirmed.
-    """
-    if context and CODE_CONTEXT_REDACTED_VALUE not in context:
-        _logger.warning(
-            "code_context passed to report without '%s' marker — "
-            "stripping content to prevent PHI leaking into report output",
-            CODE_CONTEXT_REDACTED_VALUE,
-        )
-        return CODE_CONTEXT_REDACTED_VALUE
-    return context[:_MAX_CONTEXT_CHARS]
-
-
-# ---------------------------------------------------------------------------
-# Chart generation — all charts return a matplotlib Figure
-# ---------------------------------------------------------------------------
-
-
-def _truncate_chart_label(label: str) -> str:
-    """Shorten a file path label to _MAX_LABEL_CHARS for chart readability.
-
-    Args:
-        label: Raw file path string.
-
-    Returns:
-        Original label when short enough; otherwise a "..."-prefixed tail.
-    """
-    if len(label) <= _MAX_LABEL_CHARS:
-        return label
-    return "..." + label[-(_MAX_LABEL_CHARS - 1) :]
-
-
-@dataclass(frozen=True)
-class _HorizontalBarChartSpec:
-    """Specification for rendering a horizontal bar chart figure."""
-
-    labels: tuple[str, ...]
-    values: tuple[int, ...]
-    colours: tuple[str, ...]
-    title: str
-    xlabel: str
-    chart_height_inches: float
-
-
-def _render_horizontal_bar_figure(spec: _HorizontalBarChartSpec) -> _MatplotlibFigure:
-    """Build and return a horizontal bar chart Figure from a spec."""
-    import matplotlib.pyplot as plt
-
-    figure, chart_axes = plt.subplots(figsize=(_CHART_WIDTH_INCHES, spec.chart_height_inches))
-    bars = chart_axes.barh(spec.labels, spec.values, color=spec.colours)
-    chart_axes.bar_label(
-        bars, padding=_CHART_BAR_LABEL_PADDING, fontsize=_CHART_BAR_LABEL_FONT_SIZE
-    )
-    chart_axes.set_xlabel(spec.xlabel)
-    chart_axes.set_title(spec.title)
-    chart_axes.invert_yaxis()
-    figure.tight_layout()
-    plt.close(figure)
-    return figure  # type: ignore[return-value]
-
-
-def _prepare_category_chart_data(
-    scan_result: ScanResult,
-) -> tuple[list[str], list[int], list[str]]:
-    """Return (labels, values, colours) for the HIPAA category bar chart."""
-    category_counts = {
-        cat.value.replace("_", " ").title(): count
-        for cat, count in scan_result.category_counts.items()
-        if count > 0
-    }
-    if not category_counts:
-        return [_CHART_NO_FINDINGS_LABEL], [0], [_CHART_COLOUR_ZERO_FINDINGS_BAR]
-    sorted_items = sorted(category_counts.items(), key=itemgetter(1), reverse=True)
-    labels = [label for label, _ in sorted_items]
-    finding_counts = [finding_count for _, finding_count in sorted_items]
-    colours = [
-        _CHART_COLOUR_ACTIVE if finding_count > 0 else _CHART_COLOUR_ZERO_FINDINGS_BAR
-        for finding_count in finding_counts
-    ]
-    return labels, finding_counts, colours
-
-
-def _build_category_chart(scan_result: ScanResult) -> _MatplotlibFigure:
-    """Horizontal bar chart — findings count by HIPAA category, sorted descending."""
-    labels, finding_counts, colours = _prepare_category_chart_data(scan_result)
-    return _render_horizontal_bar_figure(
-        _HorizontalBarChartSpec(
-            labels=tuple(labels),
-            values=tuple(finding_counts),
-            colours=tuple(colours),
-            title="Findings by HIPAA Category",
-            xlabel="Findings Count",
-            chart_height_inches=_CHART_HEIGHT_CATEGORY_INCHES,
-        )
-    )
-
-
-def _prepare_severity_chart_data(
-    scan_result: ScanResult,
-) -> tuple[list[str], list[int], list[str]] | None:
-    """Return (labels, values, colours) for the severity donut, or None if no findings."""
-    severity_data = {
-        level: count for level, count in scan_result.severity_counts.items() if count > 0
-    }
-    if not severity_data:
-        return None
-    ordered_levels = sorted(
-        severity_data.keys(), key=lambda severity_level: SEVERITY_RANK[severity_level], reverse=True
-    )
-    level_colours: dict[SeverityLevel, str] = {
-        SeverityLevel.HIGH: _CHART_COLOUR_SEVERITY_HIGH,
-        SeverityLevel.MEDIUM: _CHART_COLOUR_SEVERITY_MEDIUM,
-        SeverityLevel.LOW: _CHART_COLOUR_SEVERITY_LOW,
-        SeverityLevel.INFO: _CHART_COLOUR_SEVERITY_INFO,
-    }
-    labels = [f"{level.value.title()} ({severity_data[level]})" for level in ordered_levels]
-    severity_finding_counts = [severity_data[level] for level in ordered_levels]
-    colours = [level_colours.get(level, _CHART_COLOUR_UNKNOWN_SEVERITY) for level in ordered_levels]
-    return labels, severity_finding_counts, colours
-
-
-def _build_severity_chart(scan_result: ScanResult) -> _MatplotlibFigure:
-    """Donut chart — severity distribution."""
-    import matplotlib.pyplot as plt
-
-    figure, chart_axes = plt.subplots(figsize=(_CHART_WIDTH_INCHES, _CHART_HEIGHT_PIE_INCHES))
-    chart_data = _prepare_severity_chart_data(scan_result)
-    if chart_data is None:
-        chart_axes.text(
-            0.5,
-            0.5,
-            _CHART_NO_FINDINGS_TEXT,
-            ha="center",
-            va="center",
-            fontsize=_CHART_NO_DATA_FONT_SIZE,
-        )
-        chart_axes.axis("off")
-        figure.tight_layout()
-        plt.close(figure)
-        return figure  # type: ignore[return-value]
-    labels, severity_finding_counts, colours = chart_data
-    chart_axes.pie(
-        severity_finding_counts,
-        labels=labels,
-        colors=colours,
-        wedgeprops={"width": _CHART_DONUT_WIDTH},
-        startangle=_CHART_PIE_START_ANGLE,
-        autopct="%1.0f%%",
-    )
-    chart_axes.set_title("Severity Distribution")
-    figure.tight_layout()
-    plt.close(figure)
-    return figure  # type: ignore[return-value]
-
-
-def _prepare_top_files_chart_data(scan_result: ScanResult) -> tuple[list[str], list[int]]:
-    """Return (display_labels, counts) for the top-N files bar chart."""
-    file_counts: Counter[str] = Counter(str(finding.file_path) for finding in scan_result.findings)
-    top_files = file_counts.most_common(_TOP_FILES_COUNT)
-    labels = [_truncate_chart_label(file_path) for file_path, _ in top_files]
-    counts = [count for _, count in top_files]
-    return labels, counts
-
-
-def _build_top_files_chart(scan_result: ScanResult) -> _MatplotlibFigure:
-    """Horizontal bar chart — top N files with most findings."""
-    import matplotlib.pyplot as plt
-
-    labels, counts = _prepare_top_files_chart_data(scan_result)
-    figure, chart_axes = plt.subplots(figsize=(_CHART_WIDTH_INCHES, _CHART_HEIGHT_FILES_INCHES))
-    if not labels:
-        chart_axes.text(
-            0.5,
-            0.5,
-            _CHART_NO_FINDINGS_TEXT,
-            ha="center",
-            va="center",
-            fontsize=_CHART_NO_DATA_FONT_SIZE,
-        )
-        chart_axes.axis("off")
-        figure.tight_layout()
-        plt.close(figure)
-        return figure  # type: ignore[return-value]
-    bars = chart_axes.barh(labels, counts, color=_CHART_COLOUR_BLUE)
-    chart_axes.bar_label(
-        bars, padding=_CHART_BAR_LABEL_PADDING, fontsize=_CHART_BAR_LABEL_FONT_SIZE
-    )
-    chart_axes.set_xlabel("Findings Count")
-    chart_axes.set_title(f"Top {len(labels)} Files by Finding Count")
-    chart_axes.invert_yaxis()
-    figure.tight_layout()
-    plt.close(figure)
-    return figure  # type: ignore[return-value]
-
-
-def _extract_trend_data_points(
-    audit_rows: list[dict[str, object]],
-) -> tuple[_TrendDataPoint, ...]:
-    """Sanitization boundary: extract trend chart data from audit rows.
-
-    Each raw audit row is first restricted to _AUDIT_ROW_ALLOWED_KEYS via a
-    dict comprehension — this is the runtime enforcement that no other audit
-    column (repository_hash, branch_hash, findings_json, etc.) can reach
-    downstream chart rendering regardless of schema changes or unexpected fields.
-    Only the two PHI-free columns are then extracted from the restricted dict:
-    timestamp (ISO 8601 datetime string) and findings_count (integer aggregate).
-
-    Rows with an unparseable timestamp are skipped so one malformed row does not
-    suppress the entire trend chart.
-
-    Args:
-        audit_rows: Raw rows from query_recent_scans.
-
-    Returns:
-        Sanitized trend points in chronological order.
-    """
-    sorted_rows = sorted(
-        audit_rows,
-        key=lambda row: str(row.get(_AUDIT_ROW_KEY_TIMESTAMP, "")),
-    )
-    points: list[_TrendDataPoint] = []
-    for row in sorted_rows:
-        safe_row: dict[str, object] = {
-            key: row[key] for key in _AUDIT_ROW_ALLOWED_KEYS if key in row
-        }
-        raw_timestamp = safe_row.get(_AUDIT_ROW_KEY_TIMESTAMP, "")
-        try:
-            scan_date = datetime.fromisoformat(str(raw_timestamp))
-        except (ValueError, TypeError):
-            continue
-        raw_count = safe_row.get(_AUDIT_ROW_KEY_FINDINGS_COUNT, 0)
-        findings_count = int(raw_count) if isinstance(raw_count, (int, float, str)) else 0
-        points.append(_TrendDataPoint(scan_date=scan_date, findings_count=findings_count))
-    return tuple(points)
-
-
-def _build_trend_chart(trend_points: tuple[_TrendDataPoint, ...]) -> _MatplotlibFigure:
-    """Line chart — findings over time from audit log history.
-
-    Accepts only pre-sanitized _TrendDataPoint values so raw audit row dicts
-    never enter chart rendering. Callers must pass the output of
-    _extract_trend_data_points rather than raw audit_rows.
-    """
-    import matplotlib.dates as mdates
-    import matplotlib.pyplot as plt
-
-    figure, chart_axes = plt.subplots(figsize=(_CHART_WIDTH_INCHES, _CHART_HEIGHT_TREND_INCHES))
-
-    if not trend_points:
-        chart_axes.text(
-            0.5,
-            0.5,
-            _CHART_NO_HISTORY_TEXT,
-            ha="center",
-            va="center",
-            fontsize=_CHART_NO_HISTORY_FONT_SIZE,
-        )
-        chart_axes.axis("off")
-        figure.tight_layout()
-        plt.close(figure)
-        return figure  # type: ignore[return-value]
-
-    date_nums = mdates.date2num([p.scan_date for p in trend_points])  # type: ignore[no-untyped-call]
-    counts = [p.findings_count for p in trend_points]
-    chart_axes.plot(
-        date_nums,
-        counts,
-        marker=_CHART_MARKER_STYLE,
-        linewidth=_CHART_LINE_WIDTH,
-        color=_CHART_COLOUR_BLUE,
-    )
-    chart_axes.fill_between(date_nums, counts, alpha=_CHART_FILL_ALPHA, color=_CHART_COLOUR_BLUE)
-    chart_axes.xaxis.set_major_formatter(mdates.DateFormatter("%b %d"))  # type: ignore[no-untyped-call]
-    chart_axes.xaxis.set_major_locator(mdates.AutoDateLocator())  # type: ignore[no-untyped-call]
-    figure.autofmt_xdate()
-    chart_axes.set_ylabel("Findings")
-    chart_axes.set_title("Findings Trend (Audit Log History)")
-    chart_axes.grid(axis="y", linestyle="--", alpha=_CHART_GRID_ALPHA)
-    figure.tight_layout()
-    plt.close(figure)
-    return figure  # type: ignore[return-value]
-
-
-# ---------------------------------------------------------------------------
-# HTML report
-# ---------------------------------------------------------------------------
-
-_HTML_TEMPLATE: str = """\
-<!DOCTYPE html>
-<html lang="en">
-<head>
-<meta charset="UTF-8">
-<meta name="viewport" content="width=device-width, initial-scale=1.0">
-<title>PhiScan Report \u2014 {{ scan_target }}</title>
-<style>
-  :root { --font: 'Segoe UI', system-ui, -apple-system, sans-serif; }
-  * { box-sizing: border-box; margin: 0; padding: 0; }
-  body { font-family: var(--font); font-size: 14px; color: #2c3e50; background: #f8f9fa; }
-  .container { max-width: 1100px; margin: 0 auto; padding: 24px; }
-  header { background: #2c3e50; color: #fff; padding: 32px 24px; border-radius: 8px; margin-bottom: 24px; }
-  header h1 { font-size: 26px; margin-bottom: 6px; }
-  header .meta { font-size: 13px; opacity: 0.75; }
-  .risk-badge {
-    display: inline-block; padding: 6px 18px; border-radius: 20px;
-    background: #{{ risk_colour }}; color: #fff; font-weight: bold;
-    font-size: 15px; margin-top: 12px; letter-spacing: 1px; text-transform: uppercase;
-  }
-  .grid { display: grid; grid-template-columns: repeat(auto-fit, minmax(180px, 1fr)); gap: 16px; margin-bottom: 24px; }
-  .card { background: #fff; border-radius: 8px; padding: 20px; box-shadow: 0 1px 4px rgba(0,0,0,.08); }
-  .card .label { font-size: 12px; color: #7f8c8d; text-transform: uppercase; letter-spacing: .5px; margin-bottom: 6px; }
-  .card .value { font-size: 28px; font-weight: 700; color: #2c3e50; }
-  .card .sub { font-size: 12px; color: #95a5a6; margin-top: 4px; }
-  section { background: #fff; border-radius: 8px; padding: 24px; box-shadow: 0 1px 4px rgba(0,0,0,.08); margin-bottom: 24px; }
-  section h2 { font-size: 16px; font-weight: 600; margin-bottom: 16px; border-bottom: 2px solid #ecf0f1; padding-bottom: 8px; }
-  .charts-grid { display: grid; grid-template-columns: 1fr 1fr; gap: 20px; }
-  .charts-grid img { width: 100%; border-radius: 6px; border: 1px solid #ecf0f1; }
-  .full-width { grid-column: 1 / -1; }
-  table { width: 100%; border-collapse: collapse; font-size: 12px; }
-  th { background: #34495e; color: #fff; padding: 8px 10px; text-align: left; font-weight: 600; position: sticky; top: 0; }
-  td { padding: 7px 10px; border-bottom: 1px solid #ecf0f1; vertical-align: top; }
-  tr:hover td { background: #f0f4f8; }
-  .badge { display: inline-block; padding: 2px 8px; border-radius: 12px; font-size: 11px; font-weight: 600; }
-  .badge-high    { background: #fadbd8; color: #c0392b; }
-  .badge-medium  { background: #fdebd0; color: #e67e22; }
-  .badge-low     { background: #d5f5e3; color: #27ae60; }
-  .badge-info    { background: #ebf5fb; color: #2980b9; }
-  details summary { cursor: pointer; color: #2980b9; font-size: 11px; margin-top: 4px; }
-  details pre { background: #f4f6f7; padding: 8px; border-radius: 4px; font-size: 11px;
-                overflow-x: auto; margin-top: 4px; white-space: pre-wrap; word-break: break-all; }
-  .remediation-group { margin-bottom: 16px; }
-  .remediation-group h3 { font-size: 13px; font-weight: 600; margin-bottom: 4px; }
-  .remediation-group p { font-size: 12px; color: #555; line-height: 1.5; }
-  .checklist { padding-left: 20px; margin-top: 8px; }
-  .checklist li { font-size: 13px; margin-bottom: 6px; line-height: 1.4; }
-  .framework-tag { display: inline-block; margin: 1px 2px; padding: 1px 6px; border-radius: 10px;
-                   font-size: 10px; font-weight: 600; background: #eaf0fb; color: #2471a3; }
-  footer { text-align: center; font-size: 12px; color: #95a5a6; padding: 16px 0 8px; }
-  @media print {
-    body { background: #fff; }
-    section { box-shadow: none; break-inside: avoid; }
-    details, details pre { display: block; }
-  }
-</style>
-</head>
-<body>
-<div class="container">
-
-<header>
-  <h1>PHI/PII Scan Report</h1>
-  <div class="meta">
-    Target: {{ scan_target }} &nbsp;|&nbsp;
-    PhiScan v{{ version }} &nbsp;|&nbsp;
-    {{ timestamp }}
-  </div>
-  <div class="risk-badge">{{ risk_level }}</div>
-</header>
-
-<div class="grid">
-  <div class="card">
-    <div class="label">Total Findings</div>
-    <div class="value">{{ total_findings }}</div>
-    <div class="sub">across {{ files_with_findings }} file(s)</div>
-  </div>
-  <div class="card">
-    <div class="label">Files Scanned</div>
-    <div class="value">{{ files_scanned }}</div>
-    <div class="sub">{{ files_with_findings }} contain PHI</div>
-  </div>
-  <div class="card">
-    <div class="label">High Severity</div>
-    <div class="value" style="color:#c0392b">{{ count_high }}</div>
-    <div class="sub">requires immediate action</div>
-  </div>
-  <div class="card">
-    <div class="label">Medium Severity</div>
-    <div class="value" style="color:#e67e22">{{ count_medium }}</div>
-    <div class="sub">should be remediated</div>
-  </div>
-  <div class="card">
-    <div class="label">Scan Duration</div>
-    <div class="value">{{ scan_duration }}s</div>
-    <div class="sub">wall-clock time</div>
-  </div>
-</div>
-
-{% if charts %}
-<section>
-  <h2>Visual Summary</h2>
-  <div class="charts-grid">
-    {% if charts.category %}
-    <div><img src="data:image/png;base64,{{ charts.category }}" alt="Findings by category"></div>
-    {% endif %}
-    {% if charts.severity %}
-    <div><img src="data:image/png;base64,{{ charts.severity }}" alt="Severity distribution"></div>
-    {% endif %}
-    {% if charts.top_files %}
-    <div class="full-width"><img src="data:image/png;base64,{{ charts.top_files }}" alt="Top files"></div>
-    {% endif %}
-    {% if charts.trend %}
-    <div class="full-width"><img src="data:image/png;base64,{{ charts.trend }}" alt="Trend"></div>
-    {% endif %}
-  </div>
-</section>
-{% endif %}
-
-<section>
-  <h2>Findings ({{ total_findings }})</h2>
-  {% if findings %}
-  <div style="overflow-x:auto">
-  <table>
-    <thead>
-      <tr>
-        <th>#</th><th>File</th><th>Line</th><th>Entity Type</th>
-        <th>HIPAA Category</th><th>Severity</th><th>Confidence</th>
-        <th>Layer</th><th>Value Hash</th>
-      </tr>
-    </thead>
-    <tbody>
-    {% for f in findings %}
-      <tr>
-        <td>{{ loop.index }}</td>
-        <td>
-          {{ f.file_path }}
-          {% if f.code_context %}
-          <details>
-            <summary>context</summary>
-            <pre>{{ truncate_code_context(f.code_context) }}</pre>
-          </details>
-          {% endif %}
-        </td>
-        <td>{{ f.line_number }}</td>
-        <td><code>{{ f.entity_type }}</code></td>
-        <td>{{ f.hipaa_category }}</td>
-        <td><span class="badge badge-{{ f.severity }}">{{ f.severity|upper }}</span></td>
-        <td>{{ "%.2f"|format(f.confidence) }}</td>
-        <td>{{ f.detection_layer }}</td>
-        <td style="font-family:monospace;font-size:10px">{{ f.value_hash[:16] }}&hellip;</td>
-      </tr>
-    {% endfor %}
-    </tbody>
-  </table>
-  </div>
-  {% else %}
-  <p style="color:#27ae60;font-weight:600">&#10003; No findings &mdash; codebase is clean.</p>
-  {% endif %}
-</section>
-
-<section>
-  <h2>Remediation Guidance</h2>
-  {% for cat, guidance in remediation_by_category.items() %}
-  <div class="remediation-group">
-    <h3>{{ cat }}</h3>
-    <p>{{ guidance }}</p>
-  </div>
-  {% endfor %}
-  <h3 style="margin-top:16px;font-size:13px;font-weight:600">General Remediation Checklist</h3>
-  <ul class="checklist">
-  {% for item in checklist %}
-    <li>{{ item }}</li>
-  {% endfor %}
-  </ul>
-</section>
-
-{% if compliance_matrix_rows %}
-<section>
-  <h2>Compliance Matrix</h2>
-  <p style="font-size:12px;color:#555;margin-bottom:12px">
-    Applicable regulatory controls per finding. HIPAA Safe Harbor is always shown;
-    additional frameworks reflect your <code>--framework</code> selection.
-  </p>
-  <div style="overflow-x:auto">
-  <table>
-    <thead>
-      <tr>
-        <th>#</th><th>File</th><th>Line</th><th>Category</th><th>Applicable Controls</th>
-      </tr>
-    </thead>
-    <tbody>
-    {% for row in compliance_matrix_rows %}
-      <tr>
-        <td>{{ row.index }}</td>
-        <td style="font-size:11px">{{ row.file_path }}</td>
-        <td>{{ row.line_number }}</td>
-        <td>{{ row.category }}</td>
-        <td>
-          {% for fw, control_ids in row.frameworks.items() %}
-            <div style="margin-bottom:3px">
-              <strong style="font-size:10px;color:#2c3e50">{{ fw }}</strong>
-              {% for cid in control_ids %}
-                <span class="framework-tag">{{ cid }}</span>
-              {% endfor %}
-            </div>
-          {% endfor %}
-        </td>
-      </tr>
-    {% endfor %}
-    </tbody>
-  </table>
-  </div>
-</section>
-{% endif %}
-
-<section>
-  <h2>Appendix &mdash; Scan Configuration</h2>
-  <table style="width:auto">
-    <tr><td style="font-weight:600;padding-right:24px;padding-bottom:4px">Scanner Version</td><td>PhiScan v{{ version }}</td></tr>
-    <tr><td style="font-weight:600;padding-right:24px;padding-bottom:4px">Scan Target</td><td>{{ scan_target }}</td></tr>
-    <tr><td style="font-weight:600;padding-right:24px;padding-bottom:4px">Timestamp</td><td>{{ timestamp }}</td></tr>
-    <tr><td style="font-weight:600;padding-right:24px;padding-bottom:4px">Risk Level</td><td>{{ risk_level }}</td></tr>
-    <tr><td style="font-weight:600;padding-right:24px;padding-bottom:4px">Files Scanned</td><td>{{ files_scanned }}</td></tr>
-    <tr><td style="font-weight:600;padding-right:24px;padding-bottom:4px">Total Findings</td><td>{{ total_findings }}</td></tr>
-    <tr><td style="font-weight:600;padding-right:24px;padding-bottom:4px">Scan Duration</td><td>{{ scan_duration }}s</td></tr>
-  </table>
-</section>
-
-<footer>
-  Generated by PhiScan v{{ version }} &mdash; HIPAA &amp; FHIR Compliant PHI/PII Scanner &mdash;
-  All scanning executed locally. No PHI was transmitted externally.
-</footer>
-
-</div>
-</body>
-</html>
-"""
-
-
-def _build_compliance_matrix_rows(
-    scan_result: ScanResult,
-    framework_annotations: Mapping[int, tuple[ComplianceControl, ...]] | None,
-) -> list[dict[str, object]]:
-    """Build pre-processed compliance matrix rows for the HTML template.
-
-    Returns one dict per finding that has at least one applicable control,
-    with controls grouped by framework for template rendering.
-    """
-    if not framework_annotations:
-        return []
-    rows: list[dict[str, object]] = []
-    for idx, finding in enumerate(scan_result.findings):
-        controls = framework_annotations.get(idx, ())
-        if not controls:
-            continue
-        by_framework: dict[str, list[str]] = {}
-        for control in controls:
-            fw_label = control.framework.value.upper()
-            by_framework.setdefault(fw_label, []).append(control.control_id)
-        rows.append(
-            {
-                "index": idx + 1,
-                "file_path": str(finding.file_path),
-                "line_number": finding.line_number,
-                "category": finding.hipaa_category.value,
-                "frameworks": by_framework,
-            }
-        )
-    return rows
-
-
-def _build_html_context(
-    scan_result: ScanResult,
-    scan_target: Path,
-    audit_rows: list[dict[str, object]],
-    framework_annotations: Mapping[int, tuple[ComplianceControl, ...]] | None = None,
-) -> dict[str, object]:
-    """Build the Jinja2 template context dict from a ScanResult."""
-    present_categories: set[PhiCategory] = {f.hipaa_category for f in scan_result.findings}
-    remediation_by_category: dict[str, str] = {
-        cat.value.replace("_", " ").title(): HIPAA_REMEDIATION_GUIDANCE[cat]
-        for cat in PhiCategory
-        if cat in present_categories
-    }
-
-    charts: dict[str, str] = {}
-    _configure_matplotlib_backend()
-    try:
-        charts["category"] = _render_chart_to_base64(_build_category_chart(scan_result))
-        charts["severity"] = _render_chart_to_base64(_build_severity_chart(scan_result))
-        charts["top_files"] = _render_chart_to_base64(_build_top_files_chart(scan_result))
-        charts["trend"] = _render_chart_to_base64(
-            _build_trend_chart(_extract_trend_data_points(audit_rows))
-        )
-    except Exception as chart_error:  # noqa: BLE001
-        _logger.warning("Chart generation failed — report will be text-only: %s", chart_error)
-        charts = {}
-
-    severity_counts = scan_result.severity_counts
-    return {
-        "version": __version__,
-        "scan_target": str(scan_target),
-        "timestamp": datetime.now(tz=UTC).strftime("%Y-%m-%d %H:%M UTC"),
-        "risk_level": scan_result.risk_level.value.upper(),
-        "risk_colour": _get_risk_colour(scan_result.risk_level),
-        "total_findings": len(scan_result.findings),
-        "files_scanned": scan_result.files_scanned,
-        "files_with_findings": scan_result.files_with_findings,
-        "scan_duration": f"{scan_result.scan_duration:.2f}",
-        "count_high": severity_counts.get(SeverityLevel.HIGH, 0),
-        "count_medium": severity_counts.get(SeverityLevel.MEDIUM, 0),
-        "count_low": severity_counts.get(SeverityLevel.LOW, 0),
-        "count_info": severity_counts.get(SeverityLevel.INFO, 0),
-        "findings": scan_result.findings,
-        "remediation_by_category": remediation_by_category,
-        "checklist": _GENERAL_REMEDIATION_CHECKLIST,
-        "charts": charts,
-        "compliance_matrix_rows": _build_compliance_matrix_rows(scan_result, framework_annotations),
-    }
-
-
-def generate_html_report(
-    scan_result: ScanResult,
-    scan_target: Path,
-    audit_rows: list[dict[str, object]] | None = None,
-    framework_annotations: Mapping[int, tuple[ComplianceControl, ...]] | None = None,
-) -> bytes:
-    """Render a self-contained HTML report from a ScanResult.
-
-    Args:
-        scan_result: Completed scan result.
-        scan_target: Directory or file that was scanned.
-        audit_rows: Optional rows from query_recent_scans for the trend chart.
-        framework_annotations: Optional per-finding compliance controls from
-            annotate_findings(); enables the compliance matrix section.
-
-    Returns:
-        UTF-8 encoded HTML bytes of the complete self-contained report.
-    """
-    from jinja2 import Environment, Undefined
-
-    env = Environment(undefined=Undefined, autoescape=True)  # noqa: S701
-    # Expose the context truncation guard as a template global so the template
-    # cannot accidentally render raw code_context without going through the guard.
-    env.globals["truncate_code_context"] = _truncate_code_context
-    template = env.from_string(_HTML_TEMPLATE)
-    context = _build_html_context(scan_result, scan_target, audit_rows or [], framework_annotations)
-    return template.render(**context).encode("utf-8")
-
-
-# ---------------------------------------------------------------------------
-# PDF helpers
-# ---------------------------------------------------------------------------
-
-
-def _pdf_set_font(pdf: object, style: str = "", size: int = _FONT_BODY) -> None:
-    """Set the PDF font to Helvetica with the given style and size."""
-    pdf.set_font("Helvetica", style=style, size=size)  # type: ignore[attr-defined]
-
-
-def _pdf_add_section_heading(pdf: object, title: str) -> None:
-    """Write a bold section heading with a dark underline bar."""
-    pdf.ln(4)  # type: ignore[attr-defined]
-    _pdf_set_font(pdf, style="B", size=_FONT_HEADING)
-    pdf.set_text_color(*_COLOUR_NAVY_RGB)  # type: ignore[attr-defined]
-    pdf.cell(0, 8, _encode_pdf_text_as_latin1(title), new_x="LMARGIN", new_y="NEXT")  # type: ignore[attr-defined]
-    pdf.set_fill_color(*_COLOUR_NAVY_RGB)  # type: ignore[attr-defined]
-    pdf.cell(0, 0.5, "", new_x="LMARGIN", new_y="NEXT", fill=True)  # type: ignore[attr-defined]
-    pdf.ln(3)  # type: ignore[attr-defined]
-
-
-def _pdf_write_cover_page(
-    pdf: object,
-    scan_result: ScanResult,
-    scan_target: Path,
-    timestamp: str,
-) -> None:
-    """Write the cover page of the PDF report."""
-    risk_hex = _get_risk_colour(scan_result.risk_level)
-    risk_rgb = _convert_hex_to_rgb(risk_hex)
-
-    pdf.add_page()  # type: ignore[attr-defined]
-    pdf.set_fill_color(*_COLOUR_NAVY_RGB)  # type: ignore[attr-defined]
-    pdf.rect(0, 0, _PAGE_WIDTH_MM, 60, style="F")  # type: ignore[attr-defined]
-
-    _pdf_set_font(pdf, style="B", size=_FONT_TITLE)
-    pdf.set_text_color(*_COLOUR_WHITE_RGB)  # type: ignore[attr-defined]
-    pdf.set_y(_PDF_COVER_TITLE_Y_MM)  # type: ignore[attr-defined]
-    pdf.cell(0, 10, "PHI / PII Scan Report", align="C", new_x="LMARGIN", new_y="NEXT")  # type: ignore[attr-defined]
-
-    _pdf_set_font(pdf, size=_FONT_BODY)
-    pdf.set_text_color(*_COLOUR_LIGHT_SILVER_RGB)  # type: ignore[attr-defined]
-    pdf.cell(  # type: ignore[attr-defined]
-        0,
-        6,
-        _encode_pdf_text_as_latin1(f"PhiScan v{__version__}  |  {timestamp}"),
-        align="C",
-        new_x="LMARGIN",
-        new_y="NEXT",
-    )
-
-    pdf.set_y(_PDF_COVER_RISK_Y_MM)  # type: ignore[attr-defined]
-    _pdf_set_font(pdf, style="B", size=_FONT_HEADING)
-    pdf.set_text_color(*risk_rgb)  # type: ignore[attr-defined]
-    risk_label = f"RISK LEVEL: {scan_result.risk_level.value.upper()}"
-    pdf.cell(  # type: ignore[attr-defined]
-        0, 10, _encode_pdf_text_as_latin1(risk_label), align="C", new_x="LMARGIN", new_y="NEXT"
-    )
-
-    pdf.ln(8)  # type: ignore[attr-defined]
-    pdf.set_text_color(*_COLOUR_NAVY_RGB)  # type: ignore[attr-defined]
-    metadata_rows: list[tuple[str, str]] = [
-        ("Scan Target", str(scan_target)),
-        ("Total Findings", str(len(scan_result.findings))),
-        ("Files Scanned", str(scan_result.files_scanned)),
-        ("Files with Findings", str(scan_result.files_with_findings)),
-        ("Scan Duration", f"{scan_result.scan_duration:.2f}s"),
-        ("Timestamp", timestamp),
-        ("Scanner Version", f"PhiScan v{__version__}"),
-    ]
-    for label, value in metadata_rows:
-        _pdf_set_font(pdf, style="B", size=_FONT_BODY)
-        pdf.cell(55, 7, _encode_pdf_text_as_latin1(label), border="B")  # type: ignore[attr-defined]
-        _pdf_set_font(pdf, size=_FONT_BODY)
-        pdf.cell(0, 7, _encode_pdf_text_as_latin1(value), border="B", new_x="LMARGIN", new_y="NEXT")  # type: ignore[attr-defined]
-
-    pdf.ln(10)  # type: ignore[attr-defined]
-    _pdf_set_font(pdf, size=_FONT_SMALL)
-    pdf.set_text_color(*_COLOUR_MUTED_GRAY_RGB)  # type: ignore[attr-defined]
-    pdf.multi_cell(  # type: ignore[attr-defined]
-        0,
-        5,
-        "This report was generated entirely locally. No PHI or PII was transmitted "
-        "to any external service. All detected values are stored as SHA-256 hashes only.",
-    )
-
-
-def _pdf_write_summary_page(
-    pdf: object,
-    scan_result: ScanResult,
-    charts: dict[str, bytes],
-) -> None:
-    """Write the executive summary page with severity/category tables and charts."""
-    pdf.add_page()  # type: ignore[attr-defined]
-    _pdf_add_section_heading(pdf, "Executive Summary")
-
-    pdf.set_text_color(*_COLOUR_NAVY_RGB)  # type: ignore[attr-defined]
-    severity_rows: list[tuple[str, int]] = [
-        ("HIGH", scan_result.severity_counts.get(SeverityLevel.HIGH, 0)),
-        ("MEDIUM", scan_result.severity_counts.get(SeverityLevel.MEDIUM, 0)),
-        ("LOW", scan_result.severity_counts.get(SeverityLevel.LOW, 0)),
-        ("INFO", scan_result.severity_counts.get(SeverityLevel.INFO, 0)),
-    ]
-    for severity_label, count in severity_rows:
-        _pdf_set_font(pdf, size=_FONT_BODY)
-        pdf.cell(50, 6, f"{severity_label}:")  # type: ignore[attr-defined]
-        _pdf_set_font(pdf, style="B", size=_FONT_BODY)
-        pdf.cell(0, 6, str(count), new_x="LMARGIN", new_y="NEXT")  # type: ignore[attr-defined]
-
-    pdf.ln(4)  # type: ignore[attr-defined]
-    _pdf_add_section_heading(pdf, "HIPAA Category Breakdown")
-    non_zero_cats = [
-        (cat.value.replace("_", " ").title(), count)
-        for cat, count in scan_result.category_counts.items()
-        if count > 0
-    ]
-    if non_zero_cats:
-        for cat_name, count in sorted(non_zero_cats, key=lambda p: p[1], reverse=True):
-            _pdf_set_font(pdf, size=_FONT_BODY)
-            pdf.cell(60, 5, _encode_pdf_text_as_latin1(cat_name))  # type: ignore[attr-defined]
-            _pdf_set_font(pdf, style="B", size=_FONT_BODY)
-            pdf.cell(0, 5, str(count), new_x="LMARGIN", new_y="NEXT")  # type: ignore[attr-defined]
-    else:
-        _pdf_set_font(pdf, size=_FONT_BODY)
-        pdf.cell(0, 5, "No findings.", new_x="LMARGIN", new_y="NEXT")  # type: ignore[attr-defined]
-
-    # Embed category and severity charts one per page
-    for chart_key in ("category", "severity"):
-        png_bytes = charts.get(chart_key)
-        if not png_bytes:
-            continue
-        pdf.add_page()  # type: ignore[attr-defined]
-        img_buffer = io.BytesIO(png_bytes)
-        try:
-            pdf.image(img_buffer, x=_MARGIN_MM, y=None, w=_CONTENT_WIDTH_MM)  # type: ignore[attr-defined]
-        except Exception as embed_error:  # noqa: BLE001
-            _logger.warning("Could not embed chart %r: %s", chart_key, embed_error)
-
-
-def _pdf_write_findings_table(pdf: object, scan_result: ScanResult) -> None:
-    """Write the paginated findings table."""
-    pdf.add_page()  # type: ignore[attr-defined]
-    _pdf_add_section_heading(pdf, f"Detailed Findings ({len(scan_result.findings)})")
-
-    if not scan_result.findings:
-        _pdf_set_font(pdf, size=_FONT_BODY)
-        pdf.set_text_color(*_COLOUR_CLEAN_GREEN_RGB)  # type: ignore[attr-defined]
-        pdf.cell(0, 8, "No findings - codebase is clean.", new_x="LMARGIN", new_y="NEXT")  # type: ignore[attr-defined]
-        return
-
-    def _write_table_header() -> None:
-        pdf.set_fill_color(*_COLOUR_DARK_BLUE_RGB)  # type: ignore[attr-defined]
-        pdf.set_text_color(*_COLOUR_WHITE_RGB)  # type: ignore[attr-defined]
-        _pdf_set_font(pdf, style="B", size=_FONT_TABLE_HEADER)
-        for column in _PDF_TABLE_COLUMNS:
-            pdf.cell(column.width_mm, _PDF_HEADER_HEIGHT, column.header, border=1, fill=True)  # type: ignore[attr-defined]
-        pdf.ln()  # type: ignore[attr-defined]
-
-    _write_table_header()
-
-    for row_index, finding in enumerate(scan_result.findings):
-        if pdf.get_y() > _PDF_PAGE_BREAK_Y_MM:  # type: ignore[attr-defined]
-            pdf.add_page()  # type: ignore[attr-defined]
-            _write_table_header()
-
-        row_rgb = _convert_hex_to_rgb(_get_severity_row_colour(finding.severity))
-        pdf.set_fill_color(*row_rgb)  # type: ignore[attr-defined]
-        pdf.set_text_color(*_COLOUR_NAVY_RGB)  # type: ignore[attr-defined]
-        _pdf_set_font(pdf, size=_FONT_TABLE_BODY)
-
-        file_str = str(finding.file_path)
-        if len(file_str) > 30:
-            file_str = "..." + file_str[-29:]
-        entity_str = (
-            finding.entity_type[:17] + "..."
-            if len(finding.entity_type) > 18
-            else finding.entity_type
-        )
-
-        row_cells: tuple[str, ...] = (
-            str(row_index + 1),
-            file_str,
-            str(finding.line_number),
-            entity_str,
-            finding.hipaa_category.value.replace("_", " ").title(),
-            finding.severity.value.upper(),
-            f"{finding.confidence:.2f}",
-            finding.detection_layer.value,
-        )
-        for cell_text, column in zip(row_cells, _PDF_TABLE_COLUMNS):
-            pdf.cell(  # type: ignore[attr-defined]
-                column.width_mm,
-                _PDF_ROW_HEIGHT,
-                _encode_pdf_text_as_latin1(cell_text),
-                border="B",
-                fill=True,
-            )
-        pdf.ln()  # type: ignore[attr-defined]
-
-
-def _pdf_write_remediation_section(pdf: object, scan_result: ScanResult) -> None:
-    """Write the remediation guidance section."""
-    pdf.add_page()  # type: ignore[attr-defined]
-    _pdf_add_section_heading(pdf, "Remediation Guidance")
-
-    present_categories = {f.hipaa_category for f in scan_result.findings}
-    pdf.set_text_color(*_COLOUR_NAVY_RGB)  # type: ignore[attr-defined]
-
-    for category in PhiCategory:
-        if category not in present_categories:
-            continue
-        _pdf_set_font(pdf, style="B", size=_FONT_SUBHEADING)
-        pdf.cell(  # type: ignore[attr-defined]
-            0,
-            6,
-            _encode_pdf_text_as_latin1(category.value.replace("_", " ").title()),
-            new_x="LMARGIN",
-            new_y="NEXT",
-        )
-        _pdf_set_font(pdf, size=_FONT_BODY)
-        pdf.multi_cell(  # type: ignore[attr-defined]
-            0,
-            5,
-            _encode_pdf_text_as_latin1(
-                textwrap.fill(HIPAA_REMEDIATION_GUIDANCE[category], width=110)
-            ),
-        )
-        pdf.ln(2)  # type: ignore[attr-defined]
-
-    pdf.ln(4)  # type: ignore[attr-defined]
-    _pdf_set_font(pdf, style="B", size=_FONT_SUBHEADING)
-    pdf.cell(0, 6, "General Remediation Checklist", new_x="LMARGIN", new_y="NEXT")  # type: ignore[attr-defined]
-    _pdf_set_font(pdf, size=_FONT_BODY)
-    for item in _GENERAL_REMEDIATION_CHECKLIST:
-        pdf.multi_cell(0, 5, _encode_pdf_text_as_latin1(textwrap.fill(f"- {item}", width=110)))  # type: ignore[attr-defined]
-        pdf.ln(1)  # type: ignore[attr-defined]
-
-
-def _pdf_write_appendix(
-    pdf: object,
-    scan_result: ScanResult,
-    scan_target: Path,
-    timestamp: str,
-) -> None:
-    """Write the appendix page with scan configuration details and HIPAA reference."""
-    pdf.add_page()  # type: ignore[attr-defined]
-    _pdf_add_section_heading(pdf, "Appendix - Scan Configuration")
-
-    pdf.set_text_color(*_COLOUR_NAVY_RGB)  # type: ignore[attr-defined]
-    appendix_rows: list[tuple[str, str]] = [
-        ("Scanner Version", f"PhiScan v{__version__}"),
-        ("Scan Target", str(scan_target)),
-        ("Timestamp", timestamp),
-        ("Risk Level", scan_result.risk_level.value.upper()),
-        ("Files Scanned", str(scan_result.files_scanned)),
-        ("Files with Findings", str(scan_result.files_with_findings)),
-        ("Total Findings", str(len(scan_result.findings))),
-        ("Scan Duration", f"{scan_result.scan_duration:.2f}s"),
-    ]
-    for label, value in appendix_rows:
-        _pdf_set_font(pdf, style="B", size=_FONT_BODY)
-        pdf.cell(60, 6, _encode_pdf_text_as_latin1(label))  # type: ignore[attr-defined]
-        _pdf_set_font(pdf, size=_FONT_BODY)
-        pdf.cell(0, 6, _encode_pdf_text_as_latin1(value), new_x="LMARGIN", new_y="NEXT")  # type: ignore[attr-defined]
-
-    pdf.ln(8)  # type: ignore[attr-defined]
-    _pdf_set_font(pdf, size=_FONT_SMALL)
-    pdf.set_text_color(*_COLOUR_MUTED_GRAY_RGB)  # type: ignore[attr-defined]
-    pdf.multi_cell(  # type: ignore[attr-defined]
-        0,
-        5,
-        "HIPAA Safe Harbor De-identification (45 CFR SS164.514(b)(2)) requires removal "
-        "of all 18 named identifier categories. This report covers all Safe Harbor "
-        "categories plus additional regulated identifiers (42 CFR Part 2 SUD records, "
-        "quasi-identifier combinations). Expert Determination (SS164.514(b)(1)) requires "
-        "a qualified statistician and cannot be certified by automated tooling alone.",
-    )
-
-
-def _pdf_write_compliance_matrix(
-    pdf: object,
-    scan_result: ScanResult,
-    framework_annotations: Mapping[int, tuple[ComplianceControl, ...]],
-) -> None:
-    """Write the compliance matrix section listing controls per finding."""
-    pdf.add_page()  # type: ignore[attr-defined]
-    _pdf_add_section_heading(pdf, "Compliance Matrix")
-
-    pdf.set_text_color(*_COLOUR_NAVY_RGB)  # type: ignore[attr-defined]
-    _pdf_set_font(pdf, size=_FONT_SMALL)
-    pdf.multi_cell(  # type: ignore[attr-defined]
-        0,
-        5,
-        "Applicable regulatory controls per finding. HIPAA Safe Harbor is always shown; "
-        "additional frameworks reflect the --framework selection.",
-    )
-    pdf.ln(4)  # type: ignore[attr-defined]
-
-    for idx, finding in enumerate(scan_result.findings):
-        controls = framework_annotations.get(idx, ())
-        if not controls:
-            continue
-        if pdf.get_y() > _PDF_PAGE_BREAK_Y_MM:  # type: ignore[attr-defined]
-            pdf.add_page()  # type: ignore[attr-defined]
-
-        _pdf_set_font(pdf, style="B", size=_FONT_BODY)
-        finding_label = (
-            f"#{idx + 1}  {finding.file_path}:{finding.line_number}"
-            f"  [{finding.hipaa_category.value}]"
-        )
-        pdf.cell(  # type: ignore[attr-defined]
-            0,
-            5,
-            _encode_pdf_text_as_latin1(finding_label[:100]),
-            new_x="LMARGIN",
-            new_y="NEXT",
-        )
-        _pdf_set_font(pdf, size=_FONT_SMALL)
-        by_framework: dict[str, list[str]] = {}
-        for control in controls:
-            fw_label = control.framework.value.upper()
-            by_framework.setdefault(fw_label, []).append(control.control_id)
-        for fw_label, control_ids in by_framework.items():
-            ids_str = ", ".join(control_ids)
-            line = f"  {fw_label}: {ids_str}"
-            pdf.cell(  # type: ignore[attr-defined]
-                0,
-                4,
-                _encode_pdf_text_as_latin1(line[:120]),
-                new_x="LMARGIN",
-                new_y="NEXT",
-            )
-        pdf.ln(2)  # type: ignore[attr-defined]
-
-
-# ---------------------------------------------------------------------------
-# Public entry points
-# ---------------------------------------------------------------------------
-
-
-def generate_pdf_report(
-    scan_result: ScanResult,
-    scan_target: Path,
-    audit_rows: list[dict[str, object]] | None = None,
-    framework_annotations: Mapping[int, tuple[ComplianceControl, ...]] | None = None,
-) -> bytes:
-    """Render a professional PDF report from a ScanResult.
-
-    Args:
-        scan_result: Completed scan result.
-        scan_target: Directory or file that was scanned.
-        audit_rows: Optional rows from query_recent_scans for the trend chart.
-        framework_annotations: Optional per-finding compliance controls from
-            annotate_findings(); enables the compliance matrix section.
-
-    Returns:
-        Raw PDF bytes of the complete multi-page report.
-    """
-    import fpdf as fpdf_module
-
-    timestamp = datetime.now(tz=UTC).strftime("%Y-%m-%d %H:%M UTC")
-
-    charts: dict[str, bytes] = {}
-    _configure_matplotlib_backend()
-    try:
-        charts["category"] = _render_chart_to_bytes(_build_category_chart(scan_result))
-        charts["severity"] = _render_chart_to_bytes(_build_severity_chart(scan_result))
-        charts["top_files"] = _render_chart_to_bytes(_build_top_files_chart(scan_result))
-        charts["trend"] = _render_chart_to_bytes(
-            _build_trend_chart(_extract_trend_data_points(audit_rows or []))
-        )
-    except Exception as chart_error:  # noqa: BLE001
-        _logger.warning("Chart generation failed — PDF will be chart-free: %s", chart_error)
-
-    pdf = fpdf_module.FPDF(orientation="P", unit="mm", format="A4")
-    pdf.set_auto_page_break(auto=True, margin=_MARGIN_MM)
-    pdf.set_margins(_MARGIN_MM, _MARGIN_MM, _MARGIN_MM)
-
-    _pdf_write_cover_page(pdf, scan_result, scan_target, timestamp)
-    _pdf_write_summary_page(pdf, scan_result, charts)
-    _pdf_write_findings_table(pdf, scan_result)
-    _pdf_write_remediation_section(pdf, scan_result)
-    if framework_annotations:
-        _pdf_write_compliance_matrix(pdf, scan_result, framework_annotations)
-    _pdf_write_appendix(pdf, scan_result, scan_target, timestamp)
-
-    return bytes(pdf.output())
